requirements.txt
flask==2.3.3
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pandas==2.1.0
numpy==1.25.2 